make philo requested changes
