+++ src/Shared/ISessionDataSerializer.cs
@@ -1,11 +1,21 @@
 ﻿using System.Web.SessionState;
 
-namespace Microsoft.Web.RedisSessionStateProvider
+/// <summary>
+/// Provides methods to serialize and deserialize session state data.
+/// </summary>
+public interface ISessionDataSerializer
 {
-    public interface ISessionDataSerializer
-    {
-        string StorageTypeName { get; }
-        byte[] Serialize(SessionStateItemCollection data);
-        SessionStateItemCollection Deserialize(byte[] data);
-    }
+    /// <summary>
+    /// Serializes the session state data.
+    /// </summary>
+    /// <param name="data">The session state data to serialize.</param>
+    /// <returns>The serialized session state data as a byte array.</returns>
+    byte[] Serialize(SessionStateItemCollection data);
+
+    /// <summary>
+    /// Deserializes the session state data.
+    /// </summary>
+    /// <param name="data">The serialized session state data as a byte array.</param>
+    /// <returns>The deserialized session state data.</returns>
+    SessionStateItemCollection Deserialize(byte[] data);
 }
